+++ src/layer/domain/webstorage/api.ts
@@ -1,2 +1,3 @@
 export { StorageChannel } from './service/channel';
 export { localStorage, sessionStorage } from '../../infrastructure/webstorage/api';
+export { fakeStorage } from './model/storage';